RELATED: RAIL-417 Add executeAfm
