+++ package.json
@@ -63,11 +63,13 @@
     "webpack-dev-server": "1.16.1"
   },
   "dependencies": {
+    "@gooddata/typings": "0.0.3",
     "es6-promise": "3.0.2",
     "fetch-cookie": "0.4.0",
     "invariant": "2.2.2",
     "isomorphic-fetch": "2.2.1",
     "lodash": "4.17.4",
-    "md5": "2.2.1"
+    "md5": "2.2.1",
+    "qs": "6.5.1"
   }
 }
